# [CINN] Performance Optimization, fix GetGroupTileInfo